A sleep fixed the race conditions for the apex cloth (for now)

+++ Olympus/GroundPlane.cpp
@@ -3,71 +3,71 @@
 
 GroundPlane::GroundPlane()
 {
-	
+    
 }
 
 GroundPlane::GroundPlane(ID3D11DeviceContext *mDevcon, ID3D11Device *mDev, GeometryGenerator *geoGen, int planeSize, int increment) : 
-	mDevcon(mDevcon), mDev(mDev), size(planeSize), inc(increment)
+    mDevcon(mDevcon), mDev(mDev), size(planeSize), inc(increment)
 {
-	cb = new PostPBuff();
-	cb->viewInvProj;
-	cb->viewPrevProj;
+    cb = new PostPBuff();
+    cb->viewInvProj;
+    cb->viewPrevProj;
 
-	CreateGeometry(geoGen);
-	SetupBuffer();
-	SetupPipeline();
-	SetupRenderTarget();
+    CreateGeometry(geoGen);
+    SetupBuffer();
+    SetupPipeline();
+    SetupRenderTarget();
 }
 
 void GroundPlane::SetupRenderTarget()
 {
-	D3D11_TEXTURE2D_DESC textureDesc;
-	HRESULT result;
-	D3D11_RENDER_TARGET_VIEW_DESC renderTargetViewDesc;
-	D3D11_SHADER_RESOURCE_VIEW_DESC shaderResourceViewDesc;
+    D3D11_TEXTURE2D_DESC textureDesc;
+    HRESULT result;
+    D3D11_RENDER_TARGET_VIEW_DESC renderTargetViewDesc;
+    D3D11_SHADER_RESOURCE_VIEW_DESC shaderResourceViewDesc;
 
 
-	// Initialize the render target texture description.
-	ZeroMemory(&textureDesc, sizeof(textureDesc));
+    // Initialize the render target texture description.
+    ZeroMemory(&textureDesc, sizeof(textureDesc));
 
-	// Setup the render target texture description.
-	textureDesc.Width = SCREEN_WIDTH;
-	textureDesc.Height = SCREEN_HEIGHT;
-	textureDesc.MipLevels = 1;
-	textureDesc.ArraySize = 1;
-	textureDesc.Format = DXGI_FORMAT_R8G8B8A8_UNORM;
-	textureDesc.SampleDesc.Count = 1;
-	textureDesc.Usage = D3D11_USAGE_DEFAULT;
-	textureDesc.BindFlags = D3D11_BIND_RENDER_TARGET | D3D11_BIND_SHADER_RESOURCE;
-	textureDesc.CPUAccessFlags = 0;
-	textureDesc.MiscFlags = 0;
+    // Setup the render target texture description.
+    textureDesc.Width = SCREEN_WIDTH;
+    textureDesc.Height = SCREEN_HEIGHT;
+    textureDesc.MipLevels = 1;
+    textureDesc.ArraySize = 1;
+    textureDesc.Format = DXGI_FORMAT_R8G8B8A8_UNORM;
+    textureDesc.SampleDesc.Count = 1;
+    textureDesc.Usage = D3D11_USAGE_DEFAULT;
+    textureDesc.BindFlags = D3D11_BIND_RENDER_TARGET | D3D11_BIND_SHADER_RESOURCE;
+    textureDesc.CPUAccessFlags = 0;
+    textureDesc.MiscFlags = 0;
 
-	// Create the render target texture.
-	result = mDev->CreateTexture2D(&textureDesc, NULL, &mTargetTexture);
+    // Create the render target texture.
+    result = mDev->CreateTexture2D(&textureDesc, NULL, &mTargetTexture);
 
-	// Setup the description of the render target view.
-	renderTargetViewDesc.Format = textureDesc.Format;
-	renderTargetViewDesc.ViewDimension = D3D11_RTV_DIMENSION_TEXTURE2D;
-	renderTargetViewDesc.Texture2D.MipSlice = 0;
+    // Setup the description of the render target view.
+    renderTargetViewDesc.Format = textureDesc.Format;
+    renderTargetViewDesc.ViewDimension = D3D11_RTV_DIMENSION_TEXTURE2D;
+    renderTargetViewDesc.Texture2D.MipSlice = 0;
 
-	// Create the render target view.
-	result = mDev->CreateRenderTargetView(mTargetTexture, &renderTargetViewDesc, &mTargetView);
+    // Create the render target view.
+    result = mDev->CreateRenderTargetView(mTargetTexture, &renderTargetViewDesc, &mTargetView);
 
 
-	// Setup the description of the shader resource view.
-	shaderResourceViewDesc.Format = textureDesc.Format;
-	shaderResourceViewDesc.ViewDimension = D3D11_SRV_DIMENSION_TEXTURE2D;
-	shaderResourceViewDesc.Texture2D.MostDetailedMip = 0;
-	shaderResourceViewDesc.Texture2D.MipLevels = 1;
+    // Setup the description of the shader resource view.
+    shaderResourceViewDesc.Format = textureDesc.Format;
+    shaderResourceViewDesc.ViewDimension = D3D11_SRV_DIMENSION_TEXTURE2D;
+    shaderResourceViewDesc.Texture2D.MostDetailedMip = 0;
+    shaderResourceViewDesc.Texture2D.MipLevels = 1;
 
-	// Create the shader resource view.
-	result = mDev->CreateShaderResourceView(mTargetTexture, &shaderResourceViewDesc, &mShaderResourceView);
+    // Create the shader resource view.
+    result = mDev->CreateShaderResourceView(mTargetTexture, &shaderResourceViewDesc, &mShaderResourceView);
 
 }
 void GroundPlane::SetupPipeline()
 {
     // load and compile the two shaders
-	ID3D10Blob *VS, *PS;
+    ID3D10Blob *VS, *PS;
     D3DX11CompileFromFile("GroundPlane.hlsl", 0, 0, "VShader", "vs_5_0", 0, 0, 0, &VS, 0, 0);
     D3DX11CompileFromFile("GroundPlane.hlsl", 0, 0, "PShader", "ps_5_0", 0, 0, 0, &PS, 0, 0);
 
@@ -94,29 +94,29 @@ void GroundPlane::SetupPipeline()
 void GroundPlane::CreateGeometry(GeometryGenerator *geoGen)
 {
 
-	GeometryGenerator::MeshData GroundPlaneData;			   // geometry for the sky box
-	geoGen->CreateGrid(size, size, inc, inc, GroundPlaneData);
+    GeometryGenerator::MeshData GroundPlaneData;			   // geometry for the sky box
+    geoGen->CreateGrid((float)size, (float)size, inc, inc, GroundPlaneData);
 
-	PosNormalTexTan temp;
-	for(size_t i = 0; i < GroundPlaneData.Vertices.size(); i++)
+    PosNormalTexTan temp;
+    for(size_t i = 0; i < GroundPlaneData.Vertices.size(); i++)
     {
-		temp.Pos		= GroundPlaneData.Vertices[i].Position;
+        temp.Pos		= GroundPlaneData.Vertices[i].Position;
         temp.Normal		= GroundPlaneData.Vertices[i].Normal;
-		temp.Tex		= GroundPlaneData.Vertices[i].TexC;
+        temp.Tex		= GroundPlaneData.Vertices[i].TexC;
         temp.TangentU	= GroundPlaneData.Vertices[i].TangentU;
-		vertices.push_back( temp );
+        vertices.push_back( temp );
     }
 
-	for(size_t i = 0; i < GroundPlaneData.Indices.size(); i++)
-	{
-		indices.push_back( GroundPlaneData.Indices[i] );
-	}
+    for(size_t i = 0; i < GroundPlaneData.Indices.size(); i++)
+    {
+        indices.push_back( GroundPlaneData.Indices[i] );
+    }
 
 }
 
 void GroundPlane::SetupBuffer()
 {
-	
+    
     // create the vertex buffer
     D3D11_BUFFER_DESC bd;
     ZeroMemory(&bd, sizeof(bd));
@@ -132,10 +132,10 @@ void GroundPlane::SetupBuffer()
     // copy the vertices into the buffer
     D3D11_MAPPED_SUBRESOURCE ms;
     mDevcon->Map(GroundPlaneVertBuffer, NULL, D3D11_MAP_WRITE_DISCARD, NULL, &ms);    // map the buffer
-	memcpy(ms.pData, &vertices[0], sizeof(PosNormalTexTan) * vertices.size());                 // copy the data
+    memcpy(ms.pData, &vertices[0], sizeof(PosNormalTexTan) * vertices.size());                 // copy the data
     mDevcon->Unmap(GroundPlaneVertBuffer, NULL);                                      // unmap the buffer
 
-	// create the index buffer
+    // create the index buffer
     bd.Usage = D3D11_USAGE_DYNAMIC;
     bd.ByteWidth = sizeof(UINT) * indices.size();    // 3 per triangle, 12 triangles
     bd.BindFlags = D3D11_BIND_INDEX_BUFFER;
@@ -145,7 +145,7 @@ void GroundPlane::SetupBuffer()
     mDev->CreateBuffer(&bd, NULL, &GroundPlaneIndBuffer);
 
     mDevcon->Map(GroundPlaneIndBuffer, NULL, D3D11_MAP_WRITE_DISCARD, NULL, &ms);      // map the buffer
-	memcpy(ms.pData, &indices[0], sizeof(UINT) * indices.size());                     // copy the data
+    memcpy(ms.pData, &indices[0], sizeof(UINT) * indices.size());                     // copy the data
     mDevcon->Unmap(GroundPlaneIndBuffer, NULL);
 
 
@@ -162,44 +162,44 @@ void GroundPlane::SetupBuffer()
 void GroundPlane::Render(ID3D11Buffer *sceneBuff, Camera *mCam, int renderType)
 {
 
-	mDevcon->VSSetShader(mVS, 0, 0);
+    mDevcon->VSSetShader(mVS, 0, 0);
     mDevcon->PSSetShader(mPS, 0, 0);
     mDevcon->IASetInputLayout(mLayout);
 
 
-	cb->farZ = mCam->GetFarZ();
-	cb->nearZ = mCam->GetNearZ();
+    cb->farZ = mCam->GetFarZ();
+    cb->nearZ = mCam->GetNearZ();
 
-	XMStoreFloat4x4( &cb->viewInvProj, mCam->ViewProj() );
-	//XMStoreFloat4x4(&cb->viewInvProj, XMMatrixInverse( &XMMatrixDeterminant( XMLoadFloat4x4( &cb->viewInvProj ) ), XMLoadFloat4x4( &cb->viewInvProj ) ) );
+    XMStoreFloat4x4( &cb->viewInvProj, mCam->ViewProj() );
+    //XMStoreFloat4x4(&cb->viewInvProj, XMMatrixInverse( &XMMatrixDeterminant( XMLoadFloat4x4( &cb->viewInvProj ) ), XMLoadFloat4x4( &cb->viewInvProj ) ) );
 
 //	XMStoreFloat4x4( &cb->viewInvProj, XMMatrixTranspose( XMLoadFloat4x4( &cb->viewInvProj ) ) );
 //	XMStoreFloat4x4( &cb->viewPrevProj, XMMatrixTranspose( XMLoadFloat4x4( &cb->viewPrevProj ) ) );
 
-	 // select which vertex buffer to display
+     // select which vertex buffer to display
     UINT stride = sizeof(PosNormalTexTan);
     UINT offset = 0;
     mDevcon->IASetVertexBuffers(0, 1, &GroundPlaneVertBuffer, &stride, &offset);
 
-	mDevcon->IASetIndexBuffer(GroundPlaneIndBuffer, DXGI_FORMAT_R32_UINT, 0);
+    mDevcon->IASetIndexBuffer(GroundPlaneIndBuffer, DXGI_FORMAT_R32_UINT, 0);
 
     // select which primtive type we are using
     mDevcon->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
 
-	XMMATRIX matTrans;
-	
-	//matTrans = XMMatrixTranslation(mCam->GetPosition().x, mCam->GetPosition().y, mCam->GetPosition().z);
-	matTrans = XMMatrixTranslation(0,0,0);
+    XMMATRIX matTrans;
+    
+    //matTrans = XMMatrixTranslation(mCam->GetPosition().x, mCam->GetPosition().y, mCam->GetPosition().z);
+    matTrans = XMMatrixTranslation(0,0,0);
 
-	mDevcon->PSSetShaderResources(0, 1, &mShaderResourceView);
-	
-	// set the new values for the constant buffer
-	//mDevcon->UpdateSubresource(sceneBuff, 0, 0, mCam->ViewProj().m , 0, 0);
-	mDevcon->PSSetConstantBuffers(1, 1, &mConstBuffer);
-	mDevcon->UpdateSubresource(mConstBuffer, 0, 0, cb, 0, 0);
+    mDevcon->PSSetShaderResources(0, 1, &mShaderResourceView);
+    
+    // set the new values for the constant buffer
+    //mDevcon->UpdateSubresource(sceneBuff, 0, 0, mCam->ViewProj().m , 0, 0);
+    mDevcon->PSSetConstantBuffers(1, 1, &mConstBuffer);
+    mDevcon->UpdateSubresource(mConstBuffer, 0, 0, cb, 0, 0);
 
-	 // draw the vertex buffer to the back buffer
+     // draw the vertex buffer to the back buffer
     mDevcon->DrawIndexed(indices.size(), 0, 0);
 
-	XMStoreFloat4x4( &cb->viewPrevProj, mCam->ViewProj() );
+    XMStoreFloat4x4( &cb->viewPrevProj, mCam->ViewProj() );
 }
