+++ Olympus/ZeusRenderResources.cpp
@@ -49,9 +49,9 @@ bool ZeusVertexBuffer::getInteropResourceHandle(CUgraphicsResource& handle)
 void ZeusVertexBuffer::writeBuffer(const physx::NxApexRenderVertexBufferData& data, physx::PxU32 firstVertex, physx::PxU32 numVertices)
 {
     if (!mVertexBuffer || !numVertices)
-	{
-		return;
-	}
+    {
+        return;
+    }
     
     D3D11_MAPPED_SUBRESOURCE mappedResource;
     HRESULT result;
@@ -66,9 +66,9 @@ void ZeusVertexBuffer::writeBuffer(const physx::NxApexRenderVertexBufferData& da
     physx::PxU32 numSem = 0;
 
     for (physx::PxU32 i = 0; i < physx::apex::NxRenderVertexSemantic::NUM_SEMANTICS; i++)
-	{
-		physx::apex::NxRenderVertexSemantic::Enum semantic = (physx::apex::NxRenderVertexSemantic::Enum)i;
-		const physx::apex::NxApexRenderSemanticData& semanticData = data.getSemanticData(semantic);
+    {
+        physx::apex::NxRenderVertexSemantic::Enum semantic = (physx::apex::NxRenderVertexSemantic::Enum)i;
+        const physx::apex::NxApexRenderSemanticData& semanticData = data.getSemanticData(semantic);
         if (semantic == physx::apex::NxRenderVertexSemantic::POSITION || semantic == physx::apex::NxRenderVertexSemantic::TEXCOORD0 /*|| semantic == physx::apex::NxRenderVertexSemantic::NORMAL*/)
         {
             if (semanticData.data)
@@ -103,9 +103,9 @@ void ZeusVertexBuffer::writeBuffer(const physx::NxApexRenderVertexBufferData& da
             } 
         }
 
-	}
-    	
-	// Unlock the vertex buffer.
+    }
+        
+    // Unlock the vertex buffer.
     mDevcon->Unmap(mVertexBuffer, 0);
 }
 
@@ -148,9 +148,9 @@ bool ZeusIndexBuffer::getInteropResourceHandle(CUgraphicsResource& handle)
 void ZeusIndexBuffer::writeBuffer(const void* srcData, physx::PxU32 srcStride, physx::PxU32 firstDestElement, physx::PxU32 numElements)
 {
    if (!mIndexBuffer || !numElements)
-	{
-		return;
-	}
+    {
+        return;
+    }
     
     D3D11_MAPPED_SUBRESOURCE mappedResource;
     HRESULT result;
@@ -162,27 +162,27 @@ void ZeusIndexBuffer::writeBuffer(const void* srcData, physx::PxU32 srcStride, p
         return;
     }
         
-	if (srcData)
-	{
-		void* dstData = mappedResource.pData;
-		void* dstDataCopy = dstData;
+    if (srcData)
+    {
+        void* dstData = mappedResource.pData;
+        void* dstDataCopy = dstData;
 
-		PX_ASSERT(dstData);
-		if (dstData && srcStride)
-		{
-			dstData = ((physx::PxU8*)dstData) + firstDestElement * srcStride;
+        PX_ASSERT(dstData);
+        if (dstData && srcStride)
+        {
+            dstData = ((physx::PxU8*)dstData) + firstDestElement * srcStride;
             physx::PxU32 formatSize = mStride;
-			for (physx::PxU32 j = 0; j < numElements; j++)
-			{
-				memcpy(dstData, srcData, formatSize);
-				srcData = ((physx::PxU8*)srcData) + srcStride;
-				dstData = ((physx::PxU8*)dstData) + srcStride;
-			}
-		}
-	}
-	
+            for (physx::PxU32 j = 0; j < numElements; j++)
+            {
+                memcpy(dstData, srcData, formatSize);
+                srcData = ((physx::PxU8*)srcData) + srcStride;
+                dstData = ((physx::PxU8*)dstData) + srcStride;
+            }
+        }
+    }
+    
 
-	// Unlock the index buffer.
+    // Unlock the index buffer.
     mDevcon->Unmap(mIndexBuffer, 0);
 }
 
@@ -298,18 +298,18 @@ ZeusSpriteBuffer::ZeusSpriteBuffer(const physx::apex::NxUserRenderSpriteBufferDe
     for (physx::PxU32 i = 0; i < physx::apex::NxRenderSpriteSemantic::NUM_SEMANTICS; i++)
     {
         physx::apex::NxRenderSpriteSemantic::Enum apexSemantic = physx::apex::NxRenderSpriteSemantic::Enum(i);
-		physx::apex::NxRenderDataFormat::Enum apexFormat = desc.semanticFormats[i];
+        physx::apex::NxRenderDataFormat::Enum apexFormat = desc.semanticFormats[i];
      
         if (apexFormat != physx::apex::NxRenderDataFormat::UNSPECIFIED)
         {
-			if(apexSemantic == physx::apex::NxRenderSpriteSemantic::POSITION) // For right now only doing position
-				mStride += physx::apex::NxRenderDataFormat::getFormatDataSize(apexFormat);
+            if(apexSemantic == physx::apex::NxRenderSpriteSemantic::POSITION) // For right now only doing position
+                mStride += physx::apex::NxRenderDataFormat::getFormatDataSize(apexFormat);
         }
     }
 
     D3D11_BUFFER_DESC d3ddesc;
     d3ddesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
-	d3ddesc.ByteWidth = desc.maxSprites * mStride;
+    d3ddesc.ByteWidth = desc.maxSprites * mStride;
     d3ddesc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
     d3ddesc.MiscFlags = 0;
 
@@ -333,10 +333,10 @@ ZeusSpriteBuffer::ZeusSpriteBuffer(const physx::apex::NxUserRenderSpriteBufferDe
 
 ZeusSpriteBuffer::~ZeusSpriteBuffer(void)
 {
-	if(mSpriteBuffer)
-	{
-		mSpriteBuffer->Release();
-	}
+    if(mSpriteBuffer)
+    {
+        mSpriteBuffer->Release();
+    }
 }
 
 bool ZeusSpriteBuffer::getInteropResourceHandle(CUgraphicsResource& handle)
@@ -346,20 +346,20 @@ bool ZeusSpriteBuffer::getInteropResourceHandle(CUgraphicsResource& handle)
 
 void ZeusSpriteBuffer::Render(int start, int count)
 {
-	UINT stride = (UINT)mStride;
-	UINT offset = 0;
-	mDevcon->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_POINTLIST);
-	mDevcon->IASetVertexBuffers(0, 1, &mSpriteBuffer, &stride, &offset);
-	
-	mDevcon->Draw(count, start);
+    UINT stride = (UINT)mStride;
+    UINT offset = 0;
+    mDevcon->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_POINTLIST);
+    mDevcon->IASetVertexBuffers(0, 1, &mSpriteBuffer, &stride, &offset);
+    
+    mDevcon->Draw(count, start);
 }
 
 void ZeusSpriteBuffer::writeBuffer(const physx::apex::NxApexRenderSpriteBufferData& data, physx::PxU32 firstSprite, physx::PxU32 numSprites)
 {
     if (!mSpriteBuffer || !numSprites)
-	{
-		return;
-	}
+    {
+        return;
+    }
     
     D3D11_MAPPED_SUBRESOURCE mappedResource;
     HRESULT result;
@@ -372,78 +372,78 @@ void ZeusSpriteBuffer::writeBuffer(const physx::apex::NxApexRenderSpriteBufferDa
     }
 
     for (physx::PxU32 i = 0; i < physx::apex::NxRenderSpriteSemantic::NUM_SEMANTICS; i++)
-	{
-		physx::apex::NxRenderSpriteSemantic::Enum semantic = (physx::apex::NxRenderSpriteSemantic::Enum)i;
-		const physx::apex::NxApexRenderSemanticData& semanticData = data.getSemanticData(semantic);
-		if (semanticData.data)
-		{
-			const void* srcData = semanticData.data;
-			const physx::PxU32 srcStride = semanticData.stride;
+    {
+        physx::apex::NxRenderSpriteSemantic::Enum semantic = (physx::apex::NxRenderSpriteSemantic::Enum)i;
+        const physx::apex::NxApexRenderSemanticData& semanticData = data.getSemanticData(semantic);
+        if (semanticData.data)
+        {
+            const void* srcData = semanticData.data;
+            const physx::PxU32 srcStride = semanticData.stride;
 
-			physx::apex::NxRenderVertexSemantic::Enum apexSemantic = physx::apex::NxRenderVertexSemantic::NUM_SEMANTICS;
+            physx::apex::NxRenderVertexSemantic::Enum apexSemantic = physx::apex::NxRenderVertexSemantic::NUM_SEMANTICS;
             if(semantic == physx::apex::NxRenderSpriteSemantic::POSITION)
             {
                 physx::PxU32 semanticStride = mStride;
-				void* dstData = mappedResource.pData;
-				void* dstDataCopy = dstData;
-				PX_ASSERT(dstData && semanticStride);
-				if (dstData && semanticStride)
-				{
-					dstData = ((physx::PxU8*)dstData) + firstSprite * semanticStride;
+                void* dstData = mappedResource.pData;
+                void* dstDataCopy = dstData;
+                PX_ASSERT(dstData && semanticStride);
+                if (dstData && semanticStride)
+                {
+                    dstData = ((physx::PxU8*)dstData) + firstSprite * semanticStride;
                     physx::PxU32 formatSize = mStride;
-					for (physx::PxU32 j = 0; j < numSprites; j++)
-					{
-						memcpy(dstData, srcData, formatSize);
-						srcData = ((physx::PxU8*)srcData) + srcStride;
-						dstData = ((physx::PxU8*)dstData) + semanticStride;
-					}
-				}
+                    for (physx::PxU32 j = 0; j < numSprites; j++)
+                    {
+                        memcpy(dstData, srcData, formatSize);
+                        srcData = ((physx::PxU8*)srcData) + srcStride;
+                        dstData = ((physx::PxU8*)dstData) + semanticStride;
+                    }
+                }
             }
-		}
-	}
+        }
+    }
 
  //   // Get a pointer to the data in the vertex buffer.
 
  //   //const int num = &numSprites;
  //   // Copy the data into the vertex buffer.
-	//VertexType* vrticesPtr;
+    //VertexType* vrticesPtr;
  //   vrticesPtr = (VertexType*)mappedResource.pData + (firstSprite * mStride);
  //   VertexType* sauce = (VertexType*) malloc(mStride * 16);
  //   VertexType* v = new VertexType();
  //   for(physx::PxU32 i = 0; i < 16; i++)
  //   {
-	//	v->x = 0.0f;
-	//	v->y = 2.0f - (i * .125);
-	//	v->z = 0.0f;
-	//	
-	//	//memcpy( sorcData + (sizeof(v) * i), (void*)v, sizeof(v) );
-	//	//memcpy( sauce + (mStride * i), (void*)v, (mStride) );
+    //	v->x = 0.0f;
+    //	v->y = 2.0f - (i * .125);
+    //	v->z = 0.0f;
+    //	
+    //	//memcpy( sorcData + (sizeof(v) * i), (void*)v, sizeof(v) );
+    //	//memcpy( sauce + (mStride * i), (void*)v, (mStride) );
  //       for (physx::PxU32 j = 0; j < physx::apex::NxRenderSpriteSemantic::NUM_SEMANTICS; j++)
  //       {
  //           physx::apex::NxRenderSpriteSemantic::Enum apexSemantic = (physx::apex::NxRenderSpriteSemantic::Enum)j;
  //           const physx::apex::NxApexRenderSemanticData& semanticData = data.getSemanticData(apexSemantic);
  //           if (semanticData.data)
  //           {
-	//			if(apexSemantic == physx::apex::NxRenderSpriteSemantic::POSITION)
-	//			{
-	//				//memcpy(srcData + (mStride * i), semanticData.data, semanticData.stride);
-	//				VertexType* srcPtr;
-	//				srcPtr = (VertexType*)semanticData.data + (mStride * i);
-	//				memcpy(vrticesPtr + (mStride * i), (void*)srcPtr, (mStride));
-	//			}
+    //			if(apexSemantic == physx::apex::NxRenderSpriteSemantic::POSITION)
+    //			{
+    //				//memcpy(srcData + (mStride * i), semanticData.data, semanticData.stride);
+    //				VertexType* srcPtr;
+    //				srcPtr = (VertexType*)semanticData.data + (mStride * i);
+    //				memcpy(vrticesPtr + (mStride * i), (void*)srcPtr, (mStride));
+    //			}
  //           }
  //       }
  //   }
 
  //   //memcpy(vrticesPtr, (void*)sauce, (mStride));
-	
-	// Unlock the vertex buffer.
+    
+    // Unlock the vertex buffer.
     mDevcon->Unmap(mSpriteBuffer, 0);
 
-	//free(vrticesPtr);
-	/*vrticesPtr = 0;
-	delete(v);
-	v = 0;*/
+    //free(vrticesPtr);
+    /*vrticesPtr = 0;
+    delete(v);
+    v = 0;*/
 }
 
 
@@ -483,13 +483,13 @@ ZeusRenderResource::~ZeusRenderResource()
 void ZeusRenderResource::setVertexBufferRange(physx::PxU32 firstVertex, physx::PxU32 numVerts)
 {
     mVertexStart = firstVertex;
-	mVertexCount = numVerts;
+    mVertexCount = numVerts;
 }
 
 void ZeusRenderResource::setIndexBufferRange(physx::PxU32 firstIndex, physx::PxU32 numIndices)
 {
     mIndexStart = firstIndex;
-	mIndexCount = numIndices;
+    mIndexCount = numIndices;
 }
 
 void ZeusRenderResource::setBoneBufferRange(physx::PxU32 firstBone, physx::PxU32 numBones)
@@ -504,8 +504,8 @@ void ZeusRenderResource::setInstanceBufferRange(physx::PxU32 firstInstance, phys
 
 void ZeusRenderResource::setSpriteBufferRange(physx::PxU32 firstSprite, physx::PxU32 numSprites)
 {
-	mSpriteStart = firstSprite;
-	mSpriteCount = numSprites;
+    mSpriteStart = firstSprite;
+    mSpriteCount = numSprites;
 }
 
 void ZeusRenderResource::setMaterial(void* material)
@@ -515,10 +515,10 @@ void ZeusRenderResource::setMaterial(void* material)
 
 void ZeusRenderResource::Render()
 {
-	if(mSpriteBuffer)
-	{
-		mSpriteBuffer->Render(mSpriteStart, mSpriteCount);
-	}
+    if(mSpriteBuffer)
+    {
+        mSpriteBuffer->Render(mSpriteStart, mSpriteCount);
+    }
     else if(mIndexBuffer && mVertexBuffers[0])
     {
         //Render somehow?
@@ -526,16 +526,16 @@ void ZeusRenderResource::Render()
         UINT* offsets = (UINT*) malloc(sizeof(UINT) * mVertexBuffers.size());
         ID3D11Buffer* buffers = (ID3D11Buffer*) malloc(sizeof(ID3D11Buffer) * mVertexBuffers.size());
         ID3D11Buffer* buffers2[2];
-		UINT strides2[2];
-		strides2[0] = 8;
-		strides2[1] = 12;
+        UINT strides2[2];
+        strides2[0] = 8;
+        strides2[1] = 12;
         UINT offsets2[2];
-		offsets2[0] = offsets2[1] = 0;
+        offsets2[0] = offsets2[1] = 0;
 
         buffers2[0] = mVertexBuffers[0]->mVertexBuffer;
         buffers2[1] = mVertexBuffers[1]->mVertexBuffer;
 
-        for(int i = 0; i < mVertexBuffers.size(); i++)
+        for(int i = 0; i < (int)mVertexBuffers.size(); i++)
         {
             strides[i] = mVertexBuffers[i]->mStride;
             offsets[i] = 0;
@@ -545,7 +545,7 @@ void ZeusRenderResource::Render()
         mIndexBuffer->mDevcon->IASetIndexBuffer(mIndexBuffer->mIndexBuffer, DXGI_FORMAT_R32_UINT, offsets[0]);
         
         UINT stride = (UINT)12;
-	    UINT offset = 0;
+        UINT offset = 0;
         mIndexBuffer->mDevcon->IASetVertexBuffers(0, 2, buffers2, strides, offsets);
         
         mIndexBuffer->mDevcon->DrawIndexed(mIndexCount, mIndexStart, mVertexStart);
@@ -566,9 +566,9 @@ void ZeusRenderer::renderResource(const physx::apex::NxApexRenderContext& contex
 {
     if (context.renderResource)
     {
-		
+        
         //static_cast<SampleApexRendererMesh*>(context.renderResource)->render(context, mForceWireframe, mOverrideMaterial);
         //Render it here
-		static_cast<ZeusRenderResource*>(context.renderResource)->Render();
+        static_cast<ZeusRenderResource*>(context.renderResource)->Render();
     }
 }
